requirements.txt
aniso8601==3.0.2
astroid==2.2.5
autopep8==1.4.4
Django==2.1.7
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.2.1
django-heroku==0.3.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
promise==2.2.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.21.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
